track license usage for both target and ref databases
